backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
google-genai>=1.0.0
python-dotenv>=1.0.0
pillow>=10.0.0
pydantic>=2.0.0
httpx>=0.25.0
google-search-results>=2.4.2
